+++ src/clj/py99/routes/home.clj
@@ -120,7 +120,7 @@
         all-answers (db/answers-by-date)
         filter (get-in request [:session :filter] "")
         no-thanks (str/split filter #"\s")]
-    (log/debug "status-page" "no-thanks" no-thanks)
+    ;; (log/debug "status-page" "no-thanks" no-thanks)
     (layout/render
      request
      "status.html"
@@ -174,6 +174,7 @@
 
 ;; --------------------
 ;; validations
+;; use declare?
 
 (defn- remove-comments
   "Remove lines starting from #, they are comments in Python."
@@ -293,12 +294,16 @@
 (defn expand-includes
   "expand `#include` recursively."
   [s login]
-  (str/join
-   "\n"
-   (for [line (str/split-lines s)]
-     (if-let [[_ num] (re-matches #"#\s*include\s*(\d+).*" line)]
-       (expand-includes (get-answer (Integer/parseInt num) login) login)
-       line))))
+  (try
+    (str/join
+     "\n"
+     (for [line (str/split-lines s)]
+       (if-let [[_ num] (re-matches #"#\s*include\s*(\d+).*" line)]
+         (expand-includes (get-answer (Integer/parseInt num) login) login)
+         line)))
+    (catch Exception e
+      (log/error "expand-include" (.getMessage e))
+      (throw (Exception. (.getMessage e))))))
 
 (defn- has-docstring-test
   "if s contains docstring returns nil or throw.