+++ CHANGELOG.md
@@ -12,7 +12,12 @@
 - <login>-<nn> を短く <login>で表示？ 問題番号と紛らわしい。
 
 
-## v1.8-SNAPSHOT / 2024-11-09
+## v1.9-SNAPSHOT / 2024-11-12
+- fixed typo login.clj/logout
+- try expand-includes
+- (not yet)answers.doctest bool を導入するか？
+
+## v1.9-SNAPSHOT / 2024-11-12
 - コネクションプールを使い切った？
 ```
 2024-11-07 20:24:27,626 ERROR py99.middleware - nil
@@ -24,7 +29,7 @@ request timed out after 30001ms.
 - ログをもうちょっとしっかり出す。
 - layout.clj: log/debug -> log/info
 
-## v1.8-SNAPSHOT
+## v1.9-SNAPSHOT / 2024-11-12
 - login.clj: ログにリモートアドレス
 
 ## v1.7.1049 / 2024-11-04